pyproject.toml
[project]
name = "chessok"
version = "0.1.1"
description = "Multiplayer Chess Web Application with Django 6.0"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.3.2",
    "daphne>=4.2.1",
    "django==6.0.*",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.4",
    "black>=24.12.0",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "pytest-cov>=6.0.0",
    "coverage>=7.6.10",
    "django-stubs>=5.1.1",
]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
    "staticfiles",
    "media",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "I",     # isort (using standalone isort instead)
]

[tool.black]
line-length = 100
target-version = ['py313']
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | migrations
  | staticfiles
  | media
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", "migrations", "staticfiles", "media", "config/asgi.py"]
known_django = ["django"]
known_first_party = ["config", "accounts", "chess"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["."]
testpaths = ["accounts", "chess"]
addopts = [
    "--verbose",
    "--cov=.",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
    ".venv/*",
    "manage.py",
    "*/asgi.py",
    "*/wsgi.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
